Added command file support to the CLI
