+++ MinishCapRandomizerCLI/Program.cs
@@ -9,21 +9,24 @@ internal class Program
 {
     //Not Null
     private static Dictionary<string, Action> _commands;
-    //Not Null
-    private static ShufflerController _shufflerController;
-    private static string? _cachedLogicPath;
-    private static string? _cachedPatchPath;
     private static bool _exiting = false;
-    private static bool _strict = false;
-
+    
     private static void Main(string[] args)
     {
-        _shufflerController = new ShufflerController();
+        GenericCommands.ShufflerController = new ShufflerController();
         Console.Write("Loading default logic file...");
-        _shufflerController.LoadLogicFile();
+        GenericCommands.ShufflerController.LoadLogicFile();
         Console.WriteLine("done");
-        LoadCommands();
-        InputLoop();
+        if (args.Length == 1)
+        {
+            Console.WriteLine("Parsing provided commands file!");
+            CommandFileParser.ParseCommandsFile(args[0]);
+        }
+        else
+        {
+            LoadCommands();
+            InputLoop();
+        }
     }
 
     private static void InputLoop()
@@ -35,7 +38,7 @@ internal class Program
             var command = Console.ReadLine();
             if (string.IsNullOrEmpty(command) || !_commands.ContainsKey(command))
             {
-                PrintError($"Invalid command entered! {command} is not a valid command!");
+                GenericCommands.PrintError($"Invalid command entered! {command} is not a valid command!");
                 continue;
             }
             
@@ -63,7 +66,7 @@ internal class Program
             {"PatchRom", PatchRom},
             {"CreatePatch", CreatePatch},
             {"Exit", Exit},
-            {"Strict", Strict},
+            {"Strict", GenericCommands.Strict},
         };
     }
 
@@ -91,484 +94,76 @@ Strict              Toggle strict mode (exit after error)
 
     private static void LoadRom()
     {
-        Console.Write("Please enter the path to your Minish Cap Rom: ");
-        try
-        {
-            var input = Console.ReadLine();
-            if (input != null) _shufflerController.LoadRom(input);
-            Console.WriteLine("ROM Loaded Successfully!");
-        }
-        catch
-        {
-            PrintError("Failed to load ROM! Please check your file path and make sure you have read access.");
-        }
+        GenericCommands.LoadRom();
     }
 
     private static void Seed()
     {
-        Console.Write("Would you like a random seed or a set seed? (R or S): ");
-        var input = Console.ReadLine();
-        if (!string.IsNullOrEmpty(input))
-        {
-            switch (input.ToLowerInvariant())
-            {
-                case "r":
-                    var rand = new Random();
-                    _shufflerController.SetRandomizationSeed(rand.Next());
-                    Console.WriteLine("Seed set successfully!");
-                    break;
-                case "s":
-                    Console.Write("Please enter the seed you want to use: ");
-                    var seedString = Console.ReadLine();
-                    
-                    if (string.IsNullOrEmpty(seedString) || !int.TryParse(seedString, out var seed))
-                    {
-                        PrintError("Invalid seed entered!");
-                        break;
-                    }
-                    
-                    _shufflerController.SetRandomizationSeed(seed);
-                    Console.WriteLine("Seed set successfully!");
-                    break;
-                default:
-                    if (!input.Equals("exit", StringComparison.OrdinalIgnoreCase))
-                    {
-                        PrintError("Invalid Input!");
-                    }
-                    break;
-            }
-        }
-        else PrintError("Invalid Input!");
+        GenericCommands.Seed();
     }
 
     private static void LoadLogic()
     {
-        Console.Write("Please enter the path to the Logic File you want to use (leave empty to use default logic): ");
-        try
-        {
-            _cachedLogicPath = Console.ReadLine();
-            _shufflerController.LoadLogicFile(_cachedLogicPath);
-            Console.WriteLine("Logic file loaded successfully!");
-        }
-        catch
-        {
-            PrintError("Failed to load Logic File! Please check your file path and make sure you have read access.");
-        }
+        GenericCommands.LoadLogic();
     }
 
     private static void LoadPatch()
-    {        
-        Console.Write("Please enter the path to the Patch File you want to use (leave empty to use default patch): ");
-        _cachedPatchPath = Console.ReadLine();
-        Console.WriteLine("Patch file loaded successfully!");
+    {
+        GenericCommands.LoadPatch();
     }
 
     private static void LoadSettings()
     {
-        Console.Write("Please enter the setting string to load: ");
-        var input = Console.ReadLine();
-        if (!string.IsNullOrEmpty(input)) _shufflerController.LoadSettingsFromSettingString(input);
-        Console.WriteLine("Settings loaded successfully!");
+        GenericCommands.LoadSettings();
     }
 
     private static void Options()
     {
-        Console.WriteLine("Options for current logic file:");
-        var options = _shufflerController.GetLogicOptions();
-        for (var i = 0; i < options.Count; )
-        {
-            var option = options[i];
-            Console.WriteLine($"{++i}) Type: {option.GetOptionUIType()}, Option Name: {option.NiceName}, Setting Type: {option.Type}, Value: {GetOptionValue(option)}");
-        }
-
-        Console.Write("Please enter the number of the setting you would like to change, enter \"Exit\" to stop editing, or enter \"List\" to list all of the options again: ");
-        var input = Console.ReadLine();
-        
-        while (string.IsNullOrEmpty(input) || !input.Equals("exit", StringComparison.OrdinalIgnoreCase))
-        {
-            if (!string.IsNullOrEmpty(input))
-            {
-                if (input.Equals("list", StringComparison.OrdinalIgnoreCase))
-                {
-                    for (var i = 0; i < options.Count; )
-                    {
-                        var option = options[i];
-                        Console.WriteLine($"{++i}) Type: {option.GetOptionUIType()}, Option Name: {option.NiceName}, Setting Type: {option.Type}, Value: {GetOptionValue(option)}");
-                    }
-                }
-                else if (int.TryParse(input, out var num))
-                {
-                    if (--num >= 0 && num < options.Count)
-                    {
-                        EditOption(options[num]);
-                    }
-                    else
-                    {
-                        PrintError("Number ouf of range!");
-                    }
-                }
-                else
-                {
-                    if (options.Exists(option => String.Equals(option.Name, input)))
-                    {
-                        var option = options.Find(option => String.Equals(option.Name, input));
-                        if (option != null)
-                        {
-                            EditOption(option);
-                        }
-                        else
-                        {
-                            PrintError($"Unknown Option {input}!");
-                        }
-                    }
-                    else
-                    {
-                        PrintError($"Unknown Option {input}!");
-                    }
-                }
-            }
-            else
-            {
-                PrintError("Invalid Input!");
-            }
-            Console.Write("Please enter the number of the setting you would like to change, enter \"Exit\" to stop editing, or enter \"List\" to list all of the options again: ");
-            input = Console.ReadLine();
-        }
+        GenericCommands.Options();
     }
 
     private static void Logging()
     {
-        Console.WriteLine("1) Logger verbosity");
-        Console.WriteLine("2) Logger output file");
-        Console.WriteLine("3) Force publish log transactions using current verbosity and output file");
-        Console.Write("Enter the number of the logger option you want to change: ");
-        var input = Console.ReadLine();
-        if (string.IsNullOrEmpty(input) || !int.TryParse(input, out var i))
-        {
-            PrintError("Invalid Input!");
-            return;
-        }
-
-        switch (i)
-        {
-            case 1:
-            {
-                Console.WriteLine("Note: One logger transaction has many logs: info logs, warning logs, error logs, and exception logs");
-                Console.WriteLine("1) Publish all logger transactions");
-                Console.WriteLine("2) Publish only transactions that contain errors or warnings");
-                Console.Write("Enter the number for your desired verbosity: ");
-                input = Console.ReadLine();
-                if (string.IsNullOrEmpty(input) || !int.TryParse(input, out i) || (i != 1 && i != 2))
-                {
-                    PrintError("Invalid Input!");
-                    return;
-                }
-                _shufflerController.SetLoggerVerbosity(i == 1);
-                Console.WriteLine("Logger updated successfully!");
-                break;
-            }
-            case 2:
-            {
-                Console.Write("Please enter the path to save logs after randomization: ");
-                input = Console.ReadLine();
-                if (input != null) _shufflerController.SetLogOutputPath(input);
-                Console.WriteLine("Logger updated successfully!");
-                break;
-            }
-            case 3:
-                Console.WriteLine(_shufflerController.PublishLogs());
-                break;
-            default:
-                PrintError("Invalid Input!");
-                return;
-        }
+        GenericCommands.Logging();
     }
 
     private static void Randomize()
     {
-        Console.Write("How many times would you like the randomizer to attempt to generate a new seed if randomization fails? ");
-        var attemptsStr = Console.ReadLine();
-        if (!string.IsNullOrEmpty(attemptsStr) || !int.TryParse(attemptsStr, out var attempts) || attempts <= 0) attempts = 1;
-
-        _shufflerController.LoadLocations(_cachedLogicPath);
-        var success = _shufflerController.Randomize(attempts);
-        if(success)
-        {
-            Console.WriteLine("Randomization successful!");
-        }
-        else
-        {
-            PrintError("Randomization failed! " + success.ErrorMessage);
-        }
+        GenericCommands.Randomize();
     }
 
     private static void SaveRom()
-    {        
-        Console.Write("Please enter the path to save the ROM (blank for default): ");
-        try
-        {
-            var input = Console.ReadLine();
-            _shufflerController.SaveAndPatchRom(string.IsNullOrEmpty(input) ? $"{Directory.GetCurrentDirectory()}{Path.DirectorySeparatorChar}MinishRandomizer-ROM.gba" : input);
-            Console.WriteLine("Rom saved successfully!");
-        }
-        catch
-        {
-            PrintError("Failed to save ROM! Please check your file path and make sure you have write access.");
-        }
+    {
+        GenericCommands.SaveRom();
     }
 
     private static void SaveSpoiler()
-    {        
-        Console.Write("Please enter the path to save the spoiler (blank for default): ");
-        try
-        {
-            var input = Console.ReadLine();
-            _shufflerController.SaveSpoiler(string.IsNullOrEmpty(input) ? $"{Directory.GetCurrentDirectory()}{Path.DirectorySeparatorChar}MinishRandomizer-Spoiler.txt" : input);
-            Console.WriteLine("Spoiler saved successfully!");
-        }
-        catch
-        {
-            PrintError("Failed to save spoiler! Please check your file path and make sure you have write access.");
-        }
+    {
+        GenericCommands.SaveSpoiler();
     }
 
     private static void SavePatch()
     {
-        Console.Write("Please enter the path to save the patch (blank for default): ");
-        try
-        {
-            var input = Console.ReadLine();
-            _shufflerController.CreatePatch(string.IsNullOrEmpty(input) ? $"{Directory.GetCurrentDirectory()}{Path.DirectorySeparatorChar}MinishRandomizer-Patch.bps" : input, _cachedPatchPath);
-            Console.WriteLine("Patch saved successfully!");
-        }
-        catch
-        {
-            PrintError("Failed to save patch! Please check your file path and make sure you have write access.");
-        }
+        GenericCommands.SavePatch();
     }
 
     private static void GetSettingString()
     {
-        Console.WriteLine("Setting String:");
-        Console.WriteLine(_shufflerController.GetSettingsString());
+        GenericCommands.GetSettingString();
     }
 
     private static void PatchRom()
     {
-
-        Console.Write("Please enter the path of the rom patch: ");
-        var patch = Console.ReadLine();
-        
-        Console.Write("Please enter the path to save the ROM to: ");
-        var rom = Console.ReadLine();
-
-        if (rom == null || patch == null)
-        {
-            Console.WriteLine("Failed to save patch! Please check your file paths and make sure you have read/write access.");
-        }
-        else
-        {
-            var result = _shufflerController.SaveRomPatch(patch, rom);
-
-            if(result)
-            {
-                Console.WriteLine("Rom patched successfully!");
-            }
-            else
-            {
-                PrintError("Failed to patch ROM! Please check your file paths and make sure you have read/write access. "+result.ErrorMessage);
-            }
-        }
+        GenericCommands.PatchRom();
     }
 
     private static void CreatePatch()
     {
-        Console.Write("Please enter the path of the patched rom: ");
-        var rom = Console.ReadLine();
-            
-        Console.Write("Please enter the path to save the patch to: ");
-        var patch = Console.ReadLine();
-
-        if (rom == null || patch == null)
-        {
-            Console.WriteLine("Failed to save patch! Please check your file paths and make sure you have read/write access.");
-        }
-        else
-        {
-            var result = _shufflerController.SaveRomPatch(patch, rom);
-
-            if(result)
-            {
-                Console.WriteLine("Patch saved successfully!");
-            }
-            else
-            {
-                PrintError("Failed to save patch! Please check your file paths and make sure you have read/write access. "+result.ErrorMessage);
-            }
-        }
+        GenericCommands.CreatePatch();
     }
     
-    private static void Exit()
+    internal static void Exit()
     {
         _exiting = true;
-    }
-
-    private static void Strict()
-    {
-        _strict = !_strict;
-        Console.WriteLine("toggled strict mode " + (_strict ? "on" : "off"));
-    }
-
-    private static string GetOptionValue(LogicOptionBase option)
-    {
-        switch (option)
-        {
-            case LogicFlag:
-            {
-                return option.Active.ToString();
-            }
-            case LogicDropdown dropdown:
-            {
-                return dropdown.Selection.ToString();
-            }
-            case LogicColorPicker colorPicker:
-            {
-                return colorPicker.Active ? colorPicker.DefinedColor.ToString() : "Vanilla";
-            }
-            case LogicNumberBox box:
-            {
-                return box.Value;
-            }
-        }
-        return "";
-    }
-
-    private static void EditOption(LogicOptionBase option)
-    {
-        switch (option)
-        {
-            case LogicFlag:
-            {
-                Console.WriteLine("1) Enabled");
-                Console.WriteLine("2) Disabled");
-                Console.Write("Enter the number of the option to set the flag to: ");
-                var input = Console.ReadLine();
-                if (string.IsNullOrEmpty(input) || !int.TryParse(input, out var i) || (i != 0 && i != 1 && i != 2))
-                {
-                    if (!input.Equals("exit", StringComparison.OrdinalIgnoreCase))
-                    {
-                        PrintError("Invalid Input!");
-                    }
-                    break;
-                }
-
-                option.Active = i == 1;
-                Console.WriteLine("Flag set successfully!");
-                break;
-            }
-            case LogicDropdown dropdown:
-            {
-                var keys = dropdown.Selections.Keys.ToList();
-                for (var i = 0; i < keys.Count; )
-                {
-                    var selection = keys[i];
-                    Console.WriteLine($"{++i}) {selection}");
-                }
-                
-                Console.Write("Enter the number of the option you want for the dropdown: ");
-                var input = Console.ReadLine();
-                
-                if (string.IsNullOrEmpty(input) || !int.TryParse(input, out var o) || o < 1 || o > keys.Count)
-                {
-                    if (!input.Equals("exit", StringComparison.OrdinalIgnoreCase))
-                    {
-                        PrintError("Invalid Input!");
-                    }
-                    break;
-                }
-
-                dropdown.Selection = keys[o - 1];
-                Console.WriteLine("Dropdown option set successfully!");
-                break;
-            }
-            case LogicColorPicker colorPicker:
-            {
-                Console.WriteLine("1) Use Vanilla Color");
-                Console.WriteLine("2) Use Default Color");
-                Console.WriteLine("3) Use Random Color");
-                Console.WriteLine("4) Enter ARGB Color Code");
-                Console.Write("Enter the number of the option you want for the color picker: ");
-                var input = Console.ReadLine();
-                if (string.IsNullOrEmpty(input) || !int.TryParse(input, out var i) || i is < 1 or > 4)
-                {
-                    if (!input.Equals("exit", StringComparison.OrdinalIgnoreCase))
-                    {
-                        PrintError("Invalid Input!");
-                    }
-                    break;
-                }
-
-                switch (i)
-                {
-                    case 1:
-                        colorPicker.Active = false;
-                        Console.WriteLine("Color set successfully!");
-                        break;
-                    case 2:
-                        colorPicker.Active = true;
-                        colorPicker.DefinedColor = colorPicker.BaseColor;
-                        Console.WriteLine("Color set successfully!");
-                        break;
-                    case 3:
-                        colorPicker.Active = true;
-                        colorPicker.PickRandomColor();
-                        Console.WriteLine("Color set successfully!");
-                        break;
-                    case 4:
-                        Console.Write("Please enter the ARGB color string you wish to use: ");
-                        var argb = Console.ReadLine();
-                        if (string.IsNullOrEmpty(argb) || !int.TryParse(argb, out var color))
-                        {
-                            if (!argb.Equals("exit", StringComparison.OrdinalIgnoreCase))
-                            {
-                                PrintError("Invalid Input!");
-                            }
-                            break;
-                        }
-
-                        colorPicker.Active = true;
-                        colorPicker.DefinedColor = Color.FromArgb(color);
-                        Console.WriteLine("Color set successfully!");
-                        break;
-                }
-                break;
-            }
-            case LogicNumberBox box:
-            {
-                Console.Write($"Please enter a number from {box.MinValue} to {box.MaxValue} for {box.NiceName}: ");
-                var input = Console.ReadLine();
-                if (string.IsNullOrEmpty(input) || !int.TryParse(input, out var i) || i < box.MinValue || i > box.MaxValue)
-                {
-                    if (!input.Equals("exit", StringComparison.OrdinalIgnoreCase))
-                    {
-                        PrintError("Invalid Input!");
-                    }
-                    break;
-                }
-
-                box.Value = input;
-                Console.WriteLine("Number box value set successfully!");
-                break;
-            }
-        }
-    }
-
-    private static void PrintError(string msg)
-    {
-        Console.Error.WriteLine(msg);
-        if(_strict)
-        {
-            Environment.Exit(1);
-        }
     }
 }
